feat: rename conflicting file name
